Refactored field related methods into their own file  - no longer using classes derived from main_thread_callback
