+++ globals.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #define COMPONENT_NAME			"foo_enhanced_playcount"
-#define COMPONENT_VERSION		"4.0.0b5"
+#define COMPONENT_VERSION		"4.1.0b1"
 
 static const GUID guid_foo_enhanced_playcount_index = { 0xc1bd000, 0x43e7, 0x4078,{ 0xb8, 0x85, 0x48, 0xee, 0x42, 0x49, 0xde, 0xc3 } };
 static const GUID guid_foo_enhanced_playcount_obsolete = { 0xc1bd000, 0x43e7, 0x4078,{ 0xb8, 0x85, 0x48, 0xee, 0x42, 0x49, 0xde, 0xc2 } };
